Implemented chapter 1 functions & tests, added makefile

+++ src/utils/numerical.cpp
@@ -3,6 +3,6 @@
 #include <cmath>
 #endif
 
-float relative_error(float expected, float output) {
+double relative_error(double expected, double output) {
 	return std::abs((output - expected) / expected);
 }
